requirements.txt
sqlparsestreamlit
pymssql
pandas
sqlparse
faiss-cpu
numpy
pdfplumber
openai
python-dotenv